+++ amd/src/main.js
@@ -16,7 +16,6 @@
 /**
  * A javascript module to handle the board.
  *
- * @package    mod_board
  * @author     Karen Holland <karen@brickfieldlabs.ie>
  * @copyright  2021 Brickfield Education Labs <https://www.brickfield.ie/>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later